+++ Client/Writes/Events.cs
@@ -64,7 +64,11 @@ namespace InfluxDB.Client.Writes
         
         internal override void LogEvent()
         {
-            Trace.TraceError($"The retriable error occurred during writing of data. Retry in: {RetryInterval} [ms]");
+            var message = "The retriable error occurred during writing of data. " +
+                          $"Reason: '{Exception.Message}'. " +
+                          $"Retry in: {(double) RetryInterval / 1000}s.";
+            
+            Trace.TraceWarning(message);
         }
     }